Remove the frame presentation time fallback mechanism.
For the Raspberry Pi CSI2 camera being used for testing, it appears that the sensor timestamp is sometimes present and sometimes not. However, we can't mix timestamps from the system and timestamps from the sensor as they have a different offset.

It's safer not to have the timestamp until a better approach is implemented, or the root cause of missing sensor timestamps is determined and/or fixed.